+++ R/plot_group_means.R
@@ -1,50 +1,50 @@
 #' Plot group means
 #'
+#' Creates a plot of sample means and error bars by group.
+#'
 #' @param data a data object (a data frame or a data.table)
 #' @param dv_name name of the dependent variable
 #' @param iv_name name(s) of the independent variable(s).
 #' Up to two independent variables can be supplied.
-#' @param iv_name name(s) of the independent variable(s).
-#' Up to two independent variables can be supplied.
-#' @param ci if \code{TRUE}, error bars will be a confidence interval
-#' @param se if \code{TRUE}, error bars will be +/-1 standard error
-#' @param pi if \code{TRUE}, error bars will be a prediction interval
-#' @param ci_range width of the confidence interval
-#' (default = 0.95 for 95 percent confidence interval)
-#' @param pi_range width of the prediction interval
-#' (default = 0.95 for 95 percent prediction interval)
+#' @param error_bar if \code{error_bar = "se"}; error bars will be +/-1
+#' standard error, if \code{error_bar = "ci"} error bars will be a
+#' confidence interval; if \code{error_bar = "pi"}, error bars will be
+#' a prediction interval
+#' @param error_bar_range width of the confidence or prediction interval
+#' (default = 0.95 for 95 percent confidence or prediction interval).
+#' This argument will not apply when \code{error_bar = "se"}
 #' @param line_size thickness of the lines connecting group means,
 #' (default = 1)
 #' @param dot_size size of the dots indicating group means (default = 3)
-#' @param error_bar_width graphically, width of the segments
+#' @param error_bar_tip_width graphically, width of the segments
 #' at the end of error bars (default = 0.13)
 #' @param position_dodge by how much should the group means and error bars
 #' be horizontally offset from each other so as not to overlap?
 #' (default = 0.13)
 #' @param legend_position position of the legend:
-#' "none", "top", "right", "bottom", "left", "none" (default = "right")
-#' @param output output type: either "plot" or "table"
+#' \code{"none", "top", "right", "bottom", "left", "none"}
+#' (default = \code{"right"})
 #' @return by default, the output will be a ggplot object.
 #' If \code{output = "table"}, the output will be a data.table object.
 #' @examples
 #' plot_group_means(data = mtcars, dv_name = "mpg", iv_name = c("vs", "am"))
+#' plot_group_means(data = mtcars, dv_name = "mpg", iv_name = c("vs", "am"),
+#' error_bar = "se")
+#' plot_group_means(data = mtcars, dv_name = "mpg", iv_name = c("vs", "am"),
+#' error_bar = "pi", error_bar_range = 0.99)
 #' @export
 #' @import ggplot2
 plot_group_means <- function(
   data = NULL,
   dv_name = NULL,
   iv_name = NULL,
-  ci = NULL,
-  se = NULL,
-  pi = NULL,
-  ci_range = 0.95,
-  pi_range = 0.95,
+  error_bar = "ci",
+  error_bar_range = 0.95,
   line_size = 1,
   dot_size = 3,
-  error_bar_width = 0.13,
+  error_bar_tip_width = 0.13,
   position_dodge = 0.13,
-  legend_position = "right",
-  output = "plot") {
+  legend_position = "right") {
   # convert to data table
   dt1 <- data.table::setDT(data)[, c(dv_name, iv_name), with = FALSE]
   # convert iv to factors
@@ -52,31 +52,10 @@ plot_group_means <- function(
     data.table::set(dt1, j = col, value = as.factor(dt1[[col]]))
   }
   # summary table
-  dt2 <- Rmisc::summarySE(
+  dt2 <- kim::desc_stats_by_group(
     data = dt1,
-    measurevar = dv_name,
-    groupvars = iv_name,
-    conf.interval = ci_range)
-  # add 95% prediction intervals
-  dt2[["pi"]] <- dt2$sd * stats::qt(pi_range / 2 + 0.5, dt2$N - 1)
-  # output to a table
-  if (output == "table") {
-    data.table::setDT(dt2)
-    return(dt2)
-  }
-  # ci vs se vs pi
-  num_of_error_bar_args <- sum(
-    c(!is.null(ci), !is.null(se), !is.null(pi)))
-  if (num_of_error_bar_args == 0) {
-    ci <- FALSE
-    se <- FALSE
-    pi <- TRUE
-  } else if (num_of_error_bar_args >= 2) {
-    # conflicting arguments
-    stop(paste0(
-      "Please set only ONE of the following arguments to be TRUE: ",
-      "ci, se, or pi"))
-  }
+    var_for_stats = dv_name,
+    grouping_vars = iv_name)
   # check the number of ivs
   if (length(iv_name) > 2) {
     stop(paste0(
@@ -86,13 +65,13 @@ plot_group_means <- function(
   # ggplot base
   if (length(iv_name) == 1) {
     g1 <- ggplot(data = dt2, aes(
-      y = get(dv_name),
+      y = mean,
       x = get(iv_name),
       group = 1)) # connect the dots
   }
   if (length(iv_name) == 2) {
     g1 <- ggplot(data = dt2, aes(
-      y = get(dv_name),
+      y = mean,
       x = get(iv_name[1]),
       color = get(iv_name[2]),
       group = get(iv_name[2])))
@@ -101,26 +80,26 @@ plot_group_means <- function(
   # so use position_dodge to move them horizontally
   pd <- position_dodge(position_dodge)
   # build further
-  if (ci == TRUE) {
+  if (error_bar == "ci") {
     g1 <- g1 + geom_errorbar(aes(
-      ymin = get(dv_name) - ci, ymax = get(dv_name) + ci),
-      width = error_bar_width, size = line_size, position = pd)
+      ymin = dt2$ci_95_ll, ymax = dt2$ci_95_ul),
+      width = error_bar_tip_width, size = line_size, position = pd)
     error_bar_desc_text <- paste0(
-      ci_range * 100, "% confidence intervals")
+      error_bar_range * 100, "% confidence intervals")
   }
-  if (se == TRUE) {
+  if (error_bar == "se") {
     g1 <- g1 + geom_errorbar(aes(
-      ymin = get(dv_name) - se, ymax = get(dv_name) + se),
-      width = error_bar_width, size = line_size, position = pd)
+      ymin = dt2$mean - dt2$se, ymax = dt2$mean + dt2$se),
+      width = error_bar_tip_width, size = line_size, position = pd)
     error_bar_desc_text <- "one standard error (+/- 1 SE)"
   }
-  if (pi == TRUE) {
+  if (error_bar == "pi") {
     g1 <- g1 + geom_errorbar(aes(
-      ymin = get(dv_name) - pi, ymax = get(dv_name) + pi),
-      width = error_bar_width, size = line_size, position = pd)
+      ymin = dt2$pi_95_ll, ymax = dt2$pi_95_ul),
+      width = error_bar_tip_width, size = line_size, position = pd)
     error_bar_desc_text <- paste0(
-      pi_range * 100, "% prediction intervals")
-    if (pi_range == 0.95) {
+      error_bar_range * 100, "% prediction intervals")
+    if (error_bar_range == 0.95) {
       error_bar_desc_text <- paste0(
         error_bar_desc_text, " (+/- ~1.96 SD)")
     }