+++ R/se_of_mean.R
@@ -3,12 +3,14 @@
 #' @param vector a numeric vector
 #' @param na.rm if \code{TRUE}, NA values will be removed before calculation
 #' @param notify_na_count if \code{TRUE}, notify how many observations
-#' were removed due to missing values.
+#' were removed due to missing values. By default, NA count will be printed
+#' only if there are any NA values.
 #' @examples
 #' se_of_mean(c(1:10, NA))
 #' @export
 se_of_mean <- function(vector, na.rm = TRUE, notify_na_count = NULL) {
   x <- vector[!is.na(vector)]
+  # by default, notify only if NA values are present
   if (is.null(notify_na_count)) {
     notify_na_count <- ifelse(length(vector) - length(x) > 0, TRUE, FALSE)
   }